+++ packages/cloudflare/src/cli/templates/worker.ts
@@ -15,7 +15,7 @@ Object.defineProperty(globalThis, Symbol.for("__cloudflare-context__"), {
 });
 
 //@ts-expect-error: Will be resolved by wrangler build
-export { DurableObjectQueueHandler } from "./.build/durable-objects/queue.js";
+export { DOQueueHandler } from "./.build/durable-objects/queue.js";
 //@ts-expect-error: Will be resolved by wrangler build
 export { DOShardedTagCache } from "./.build/durable-objects/sharded-tag-cache.js";